refactor: 抽取 page title
